nvimのambiwidthをシングルに変更

+++ .config/nvim/init.lua
@@ -5,7 +5,7 @@ vim.o.shiftwidth = 2
 vim.o.autoindent = true
 vim.o.number = true
 vim.o.list = true
-vim.o.ambiwidth = 'double'
+vim.o.ambiwidth = 'single'
 vim.o.conceallevel = 1
 
 vim.cmd 'syntax enable'
